+++ activesupport/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Use `UNLINK` command instead of `DEL` in `ActiveSupport::Cache::RedisCacheStore` for non-blocking deletion.
+
+    *Aron Roh*
+
 *   Add `Cache#read_counter` and `Cache#write_counter`
 
     ```ruby